MCPs/graphql/requirements.txt
mcp>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
graphql-core>=3.2.0